refactor(harness): split openai provider mod.rs by concern
providers/openai/mod.rs had grown to ~1360 lines mixing HTTP
transport, request/response conversion, and SSE stream parsing. Split
mechanically into transport.rs (OpenAiModel construction, presets,
request building, the ChatModel impl), convert.rs (message/response
translation), and sse.rs (SseState/OpenAiStreamAcc/sse_next). Public
API (OpenAiModel) re-exported from mod.rs unchanged. No behavior
change.
