+++ src/harness/providers/openai/convert.rs
@@ -1,0 +1,322 @@
+//! Request/response conversion between the provider-neutral harness
+//! types and the OpenAI wire format (`translate_message`,
+//! `parse_response`, usage conversion, reasoning-text extraction).
+//!
+//! Split out of `openai/mod.rs`; see that module's doc comment for the
+//! full provider overview.
+
+use super::*;
+
+/// Translates one harness [`Message`] into an OpenAI wire message.
+///
+/// User messages are rendered as OpenAI content-parts when they carry non-text
+/// blocks (for example images), so image inputs are actually sent rather than
+/// silently dropped. Blocks that have no faithful OpenAI representation return a
+/// [`TinyAgentsError::Validation`] instead of being discarded.
+pub(super) fn translate_message(message: &Message) -> Result<ChatMessageWire> {
+    let wire = match message {
+        Message::System(_) => ChatMessageWire {
+            role: "system".to_string(),
+            content: Some(MessageContentWire::Text(message.text())),
+            tool_calls: Vec::new(),
+            tool_call_id: None,
+        },
+        Message::User(user) => ChatMessageWire {
+            role: "user".to_string(),
+            content: Some(translate_user_content(&user.content)?),
+            tool_calls: Vec::new(),
+            tool_call_id: None,
+        },
+        Message::Assistant(assistant) => {
+            let text = message.text();
+            // OpenAI accepts a null content for tool-call-only assistant turns.
+            let content = if text.is_empty() && !assistant.tool_calls.is_empty() {
+                None
+            } else {
+                Some(MessageContentWire::Text(text))
+            };
+            let tool_calls = assistant
+                .tool_calls
+                .iter()
+                .map(|call| {
+                    Ok(ToolCallWire {
+                        id: call.id.clone(),
+                        kind: "function".to_string(),
+                        function: FunctionCallWire {
+                            name: call.name.clone(),
+                            // OpenAI expects arguments as a JSON string.
+                            arguments: serde_json::to_string(&call.arguments)?,
+                        },
+                    })
+                })
+                .collect::<Result<Vec<_>>>()?;
+            ChatMessageWire {
+                role: "assistant".to_string(),
+                content,
+                tool_calls,
+                tool_call_id: None,
+            }
+        }
+        Message::Tool(tool) => ChatMessageWire {
+            role: "tool".to_string(),
+            content: Some(MessageContentWire::Text(message.text())),
+            tool_calls: Vec::new(),
+            tool_call_id: Some(tool.tool_call_id.clone()),
+        },
+    };
+    Ok(wire)
+}
+
+/// Renders user-message content blocks into OpenAI message content.
+///
+/// Text-only content collapses to a plain string (preserving the historical wire
+/// shape). When an image block is present, content is emitted as OpenAI
+/// content-parts so the image is actually sent. JSON blocks are serialized into
+/// text parts. A [`ContentBlock::ProviderExtension`] has no faithful OpenAI
+/// representation, so it fails closed with a validation error rather than being
+/// silently dropped.
+pub(super) fn translate_user_content(blocks: &[ContentBlock]) -> Result<MessageContentWire> {
+    let has_image = blocks
+        .iter()
+        .any(|block| matches!(block, ContentBlock::Image(_)));
+
+    if !has_image {
+        // No image: render as a single string, but still fail closed on blocks
+        // that cannot be represented.
+        let mut text = String::new();
+        for block in blocks {
+            match block {
+                ContentBlock::Text(t) => text.push_str(t),
+                ContentBlock::Json(value) => text.push_str(&value.to_string()),
+                ContentBlock::Image(_) => unreachable!("guarded by has_image"),
+                ContentBlock::ProviderExtension(_) => {
+                    return Err(unrepresentable_block_error());
+                }
+            }
+        }
+        return Ok(MessageContentWire::Text(text));
+    }
+
+    let mut parts = Vec::with_capacity(blocks.len());
+    for block in blocks {
+        match block {
+            ContentBlock::Text(t) => parts.push(ContentPartWire::Text { text: t.clone() }),
+            ContentBlock::Json(value) => parts.push(ContentPartWire::Text {
+                text: value.to_string(),
+            }),
+            ContentBlock::Image(image) => parts.push(ContentPartWire::ImageUrl {
+                image_url: ImageUrlWire {
+                    url: image.url.clone(),
+                },
+            }),
+            ContentBlock::ProviderExtension(_) => {
+                return Err(unrepresentable_block_error());
+            }
+        }
+    }
+    Ok(MessageContentWire::Parts(parts))
+}
+
+/// Error returned when a content block cannot be represented in an OpenAI
+/// request. Failing closed keeps the block from being silently dropped.
+pub(super) fn unrepresentable_block_error() -> TinyAgentsError {
+    TinyAgentsError::Validation(
+        "OpenAI request cannot represent a provider-extension content block; \
+         remove it or target the originating provider"
+            .to_string(),
+    )
+}
+
+/// Translates a [`ToolChoice`] into the OpenAI `tool_choice` JSON value.
+pub(super) fn translate_tool_choice(choice: &ToolChoice) -> Value {
+    match choice {
+        ToolChoice::Auto => json!("auto"),
+        ToolChoice::None => json!("none"),
+        ToolChoice::Required => json!("required"),
+        ToolChoice::Tool(name) => json!({
+            "type": "function",
+            "function": { "name": name }
+        }),
+    }
+}
+
+/// Translates a [`ResponseFormat`] into the OpenAI `response_format` JSON value.
+///
+/// Returns `None` for [`ResponseFormat::Text`] so the field is omitted entirely.
+pub(super) fn translate_response_format(format: &ResponseFormat) -> Option<Value> {
+    match format {
+        ResponseFormat::Text => None,
+        ResponseFormat::JsonObject => Some(json!({ "type": "json_object" })),
+        // OpenAI supports native structured output, so `Auto` maps to a JSON
+        // schema request directly. (The agent loop normally resolves `Auto`
+        // before reaching the provider; this keeps direct calls correct too.)
+        ResponseFormat::JsonSchema { name, schema } | ResponseFormat::Auto { name, schema } => {
+            Some(json!({
+                "type": "json_schema",
+                "json_schema": {
+                    "name": name,
+                    "schema": schema,
+                    "strict": true,
+                }
+            }))
+        }
+    }
+}
+
+/// Parses an OpenAI response body (already decoded into a [`Value`]) into a
+/// provider-neutral [`ModelResponse`].
+///
+/// The first choice is used. The raw JSON is preserved in
+/// [`ModelResponse::raw`].
+///
+/// # Errors
+///
+/// Returns [`TinyAgentsError::Serialization`] if the value does not match the
+/// expected response shape, or [`TinyAgentsError::Model`] when no choices are
+/// present.
+pub(super) fn parse_response(value: Value) -> Result<ModelResponse> {
+    let parsed: ChatCompletionResponse = serde_json::from_value(value.clone())?;
+
+    let choice = parsed.choices.into_iter().next().ok_or_else(|| {
+        TinyAgentsError::Model("openai response contained no choices".to_string())
+    })?;
+
+    let mut content = Vec::new();
+    if let Some(text) = choice.message.content.filter(|t| !t.is_empty()) {
+        content.push(ContentBlock::Text(text));
+    }
+
+    let tool_calls = choice
+        .message
+        .tool_calls
+        .into_iter()
+        .map(|call| {
+            Ok(ToolCall {
+                id: call.id.clone(),
+                name: call.function.name.clone(),
+                // Tool arguments arrive as a JSON string. Invalid JSON is a
+                // provider/model error, not an empty/default argument payload.
+                arguments: parse_tool_arguments(
+                    "openai response",
+                    &call.id,
+                    &call.function.name,
+                    &call.function.arguments,
+                )?,
+            })
+        })
+        .collect::<Result<Vec<_>>>()?;
+
+    let usage = parsed.usage.map(convert_usage);
+
+    let message = AssistantMessage {
+        id: parsed.id,
+        content,
+        tool_calls,
+        usage,
+    };
+
+    Ok(ModelResponse {
+        message,
+        usage,
+        finish_reason: choice.finish_reason,
+        raw: Some(value),
+        resolved_model: None,
+    })
+}
+
+/// Returns the effective call id for a streamed tool-call slot: the
+/// provider-assigned id when present, or a stable `tool-{slot}` fallback keyed to
+/// the slot's position so delta ids and the final call id always agree.
+pub(super) fn tool_call_id(slot: usize, id: &str) -> String {
+    if id.is_empty() {
+        format!("tool-{slot}")
+    } else {
+        id.to_string()
+    }
+}
+
+pub(super) fn parse_tool_arguments(
+    context: &str,
+    call_id: &str,
+    name: &str,
+    raw: &str,
+) -> Result<Value> {
+    // Some OpenAI-compatible backends emit an empty arguments string for a
+    // zero-argument tool call. That is a well-formed "no arguments" payload, not
+    // malformed JSON, so map it to an empty object instead of failing the call.
+    if raw.trim().is_empty() {
+        return Ok(Value::Object(Map::new()));
+    }
+    serde_json::from_str(raw).map_err(|err| {
+        TinyAgentsError::Model(format!(
+            "{context} contained invalid JSON arguments for tool call `{call_id}` (`{name}`): {err}; raw arguments: {raw:?}"
+        ))
+    })
+}
+
+/// Converts an OpenAI [`UsageWire`] into the harness-neutral [`Usage`].
+pub(super) fn convert_usage(wire: UsageWire) -> Usage {
+    // OpenAI-compatible endpoints sometimes omit `total_tokens` entirely
+    // (deserializes to `0` via `#[serde(default)]`); fall back to
+    // `prompt + completion` so `total_tokens` is never a misleading zero for
+    // a call that clearly consumed tokens.
+    let total_tokens = if wire.total_tokens > 0 {
+        wire.total_tokens
+    } else {
+        wire.prompt_tokens + wire.completion_tokens
+    };
+    Usage {
+        input_tokens: wire.prompt_tokens,
+        output_tokens: wire.completion_tokens,
+        total_tokens,
+        cache_read_tokens: wire
+            .prompt_tokens_details
+            .map(|d| d.cached_tokens)
+            .unwrap_or(0),
+        reasoning_tokens: wire
+            .completion_tokens_details
+            .map(|d| d.reasoning_tokens)
+            .unwrap_or(0),
+        ..Usage::default()
+    }
+}
+
+/// Normalizes provider-specific reasoning/thinking payloads into text.
+///
+/// OpenAI-compatible gateways do not agree on this field: some stream a plain
+/// `reasoning_content` string, others use `reasoning`, and a few wrap text in
+/// an object/array. Preserve renderable text when obvious and ignore opaque
+/// shapes rather than failing an otherwise valid completion.
+pub(super) fn reasoning_value_text(value: Value) -> Option<String> {
+    match value {
+        Value::String(text) => (!text.is_empty()).then_some(text),
+        Value::Object(map) => ["text", "content", "summary"]
+            .into_iter()
+            .find_map(|key| map.get(key).and_then(Value::as_str))
+            .filter(|text| !text.is_empty())
+            .map(str::to_string),
+        Value::Array(values) => {
+            let text = values
+                .into_iter()
+                .filter_map(reasoning_value_text)
+                .collect::<String>();
+            (!text.is_empty()).then_some(text)
+        }
+        _ => None,
+    }
+}
+
+/// Extracts the reasoning/thinking text from a streamed delta, accepting the
+/// common OpenAI-compatible aliases.
+pub(super) fn delta_reasoning_text(delta: &mut ChunkDeltaWire) -> String {
+    let mut text = String::new();
+    for value in [delta.reasoning_content.take(), delta.reasoning.take()]
+        .into_iter()
+        .flatten()
+    {
+        if let Some(fragment) = reasoning_value_text(value) {
+            text.push_str(&fragment);
+        }
+    }
+    text
+}
